revert back to vector <Player*> player for gameengine
