+++ Warzone/src/GameEngine.cpp
@@ -10,11 +10,39 @@
 
 using namespace std;
 
+// main driver
+int main() {
+    int numberOfPlayers = 0;
+    vector<Player*> tempPlayers;
+    MapLoader* maploader;
+    Map* map;
+    GameEngine* gameEngine = new GameEngine(numberOfPlayers, tempPlayers, maploader, map);
+
+    cout << "Map BEFORE user chooses: " << gameEngine->map << endl;
+    map = gameEngine->chooseMap(); // choose map to start the game
+    gameEngine->setMap(map); // set the new map from the maploader
+    cout << "Map AFTER user chooses: " << gameEngine->map << endl;
+
+    cout << "Number of player before user chooses: " << gameEngine->numberOfPlayers << endl;
+    gameEngine->setPlayerNum(); // let user choose number of players
+    cout << "Number of player after user chooses: " << gameEngine->numberOfPlayers << endl;
+
+    cout << "Creating the players and required items for the game: " << endl;
+    gameEngine->createPlayers();
+    cout << "Players have been created" << endl;
+    cout << "GameEngine player size " << gameEngine->players.size() << endl;
+
+    // free memory and dangling ptr
+    delete gameEngine;
+    gameEngine = NULL;
+
+	return 0;
+}
 
 // default constructor auto generated
 
 // constructor with params 
-GameEngine::GameEngine(int numberOfPlayers, vector<Player*> *players, MapLoader* maploader, Map* map)  {
+GameEngine::GameEngine(int numberOfPlayers, vector<Player*> players, MapLoader* maploader, Map* map)  {
        this->numberOfPlayers = numberOfPlayers;
 	   this->players = players;
 	   this->maploader = maploader;
@@ -45,7 +73,7 @@ GameEngine& GameEngine::operator=(const GameEngine &gameEngine){
 //stream insertion operator for GameEngine
 ostream& operator << (ostream &os, const GameEngine &gameEngine){ 
 	return os << "\n GameEngine number of player:" << gameEngine.numberOfPlayers 
-    << "\n GameEngine player size: " << gameEngine.numberOfPlayers  
+    << "\n GameEngine player size: " << gameEngine.players.size()  
     << "\n GameEngine maploader: " << gameEngine.maploader  
     << "\n GameEngine map: " << gameEngine.map;
 }
@@ -53,7 +81,7 @@ ostream& operator << (ostream &os, const GameEngine &gameEngine){
 // destructor
 GameEngine::~GameEngine(){
     numberOfPlayers = 0;
-    players->clear(); // empty array
+    players.clear(); // empty array
     maploader = NULL;
     map = NULL;
 }
@@ -112,7 +140,7 @@ void GameEngine::createPlayers() {
         // last parameter i is the playerID
         // create new players
         Player* tempPlayer = new Player(defaultTerritories, defaultHand, defaultOrders, i);
-        players->push_back(tempPlayer); // add the players into the GameEngine vector
+        this->players.push_back(tempPlayer); // add the players into the GameEngine vector
 
         cout << "Player " << i << " has been created " << endl;
 
@@ -147,7 +175,7 @@ void GameEngine::reinforcementPhase() {
 
         cout << "Test Round " << a << endl;
 
-    for(it = players->begin(); it != players->end(); it++){   //iterating through list of players
+    for(it = this->players.begin(); it != this->players.end(); it++){   //iterating through list of players
         // (# of territories owned divided by 3, rounded down
         cout << "Starting to loop throw each player" << endl;
         cout << (*it)->getTerritories().size() << endl;
